Fixed a bug in the case where the effect_target is set to ask, but the player-chosen target is a square. Was allowing multiple buildings on the same meadow/exit. Also fixed a potential bug in find_meadows when returning exits. It wasn't checking if the located exit was blocked.

+++ soundrts/worldroom.py
@@ -181,6 +181,7 @@ class Square(object):
         else: land_type = Meadow
         for o in self.objects:
             if isinstance(o, land_type):
+                if land_type is Exit and o.is_blocked(): continue
                 x, y = o.x, o.y
                 if isinstance(o, Meadow): o.delete()
                 return x, y, o
